Adding postspage
I made a rough html document to layout  how the post page might look.

It's similar to the index page. It uses the same style sheet so far.

I plan on changing a lot of the ID's/codenames to get something cleaner once I get time to sit down with this assignment.